day11 chapter2.

+++ Day11/Models/Monkey.cs
@@ -2,17 +2,17 @@
 {
     internal class Monkey
     {
-        public int Id { get; set; }
-        public List<int> Items { get; set; }
-        public int InspectedItemsCount { get; set; }
+        public ulong Id { get; set; }
+        public List<ulong> Items { get; set; }
+        public ulong InspectedItemsCount { get; set; }
         public char Operator { get; set; }
         public string Operant1 { get; set; }
         public string Operant2 { get; set; }
-        public int Divisor { get; set; }
-        public int MonkeyIdIfTrue { get; set; }
-        public int MonkeyIdIfFalse { get; set; }
+        public ulong Divisor { get; set; }
+        public ulong MonkeyIdIfTrue { get; set; }
+        public ulong MonkeyIdIfFalse { get; set; }
 
-        public bool Test(int worry)
+        public bool Test(ulong worry)
         {
             return worry % Divisor == 0;
         }
